# bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Periodico, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Roles/{id}": {
      "get": {
        "tags": [
          "Roles"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Rol"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Roles"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Roles/Actualizar": {
      "put": {
        "tags": [
          "Roles"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "descripcion",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Roles/Crear": {
      "post": {
        "tags": [
          "Roles"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "descripcion",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/UsuarioRoles/{id}": {
      "get": {
        "tags": [
          "UsuarioRoles"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UsuarioRol"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "UsuarioRoles"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/UsuarioRoles/Actualizar": {
      "put": {
        "tags": [
          "UsuarioRoles"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "idrol",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "idusuario",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/UsuarioRoles/Crear": {
      "post": {
        "tags": [
          "UsuarioRoles"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "idrol",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "idusuario",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Usuarios/{id}": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Usuario"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Usuarios"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Usuarios/Login": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "nombreUsuario",
            "type": "string"
          },
          {
            "in": "query",
            "name": "password",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Usuarios/Actualizar": {
      "put": {
        "tags": [
          "Usuarios"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "nombre",
            "type": "string"
          },
          {
            "in": "query",
            "name": "telefono",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "nombreusuario",
            "type": "string"
          },
          {
            "in": "query",
            "name": "password",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Usuarios/Crear": {
      "post": {
        "tags": [
          "Usuarios"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "nombre",
            "type": "string"
          },
          {
            "in": "query",
            "name": "telefono",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "nombreusuario",
            "type": "string"
          },
          {
            "in": "query",
            "name": "password",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
...
    "Detalle_Permiso": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "estado": {
          "type": "string"
        },
        "idusuario": {
          "format": "int32",
          "type": "integer"
        },
        "idpermiso": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
    "Publicacion": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "titulo": {
          "type": "string"
        },
        "fecha": {
          "format": "date-time",
          "type": "string"
        },
        "descripcion": {
          "type": "string"
        },
        "imagen": {
          "type": "string"
        },
        "visualizacion": {
          "format": "int32",
          "type": "integer"
        },
        "estado": {
          "type": "string"
        },
        "idusuario": {
          "format": "int32",
          "type": "integer"
        },
        "idcategoria": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "Rol": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "descripcion": {
          "type": "string"
        },
        "estado": {
          "type": "string"
        },
        "usuariosRoles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UsuarioRol"
          }
        }
      }
    },
...
    "Usuario": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "telefono": {
          "format": "int32",
          "type": "integer"
        },
        "nombreUsuario": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "estado": {
          "type": "string"
        },
        "publicaciones": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Publicacion"
          }
        },
        "detallePermisos": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Detalle_Permiso"
          }
        },
        "usuariosRoles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UsuarioRol"
          }
        }
      }
    },
    "UsuarioRol": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "estado": {
          "type": "string"
        },
        "idusuario": {
          "format": "int32",
          "type": "integer"
        },
        "idrol": {
          "format": "int32",
          "type": "integer"
        }
      }
    }
  }
}